+++ nodes/nomos-node-api/src/http/consensus.rs
@@ -16,10 +16,6 @@ use nomos_core::{
 use nomos_mempool::{
     backend::mockpool::MockPool, network::adapters::libp2p::Libp2pAdapter as MempoolLibp2pAdapter,
 };
-use nomos_network::{
-    backends::libp2p::{Command, Libp2p, Libp2pInfo},
-    NetworkMsg, NetworkService,
-};
 use nomos_storage::backends::{sled::SledBackend, StorageSerde};
 use serde::{de::DeserializeOwned, Serialize};
 use tokio::sync::oneshot;
@@ -56,16 +52,4 @@ where
     Ok(receiver.await?)
 }
 
-pub(crate) async fn libp2p_info(
-    handle: &overwatch_rs::overwatch::handle::OverwatchHandle,
-) -> Result<Libp2pInfo, overwatch_rs::DynError> {
-    let relay = handle.relay::<NetworkService<Libp2p>>().connect().await?;
-    let (sender, receiver) = oneshot::channel();
 
-    relay
-        .send(NetworkMsg::Process(Command::Info { reply: sender }))
-        .await
-        .map_err(|(e, _)| e)?;
-
-    Ok(receiver.await?)
-}
