+++ nodes/nomos-node-api/src/http/mod.rs
@@ -3,4 +3,5 @@ pub type DynError = Box<dyn std::error::Error + Send + Sync + 'static>;
 pub mod backend;
 pub(crate) mod cl;
 pub(crate) mod da;
-pub(crate) mod info;
+pub(crate) mod libp2p;
+pub(crate) mod consensus;
